backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
# mysqlclient==2.2.0
Pillow==10.1.0
reportlab==4.0.9
openpyxl==3.1.2
pandas==2.1.3
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
qrcode==7.4.2
pdf2image==1.16.3
PyPDF2==3.0.1
django-cleanup==8.0.0
django-import-export==3.3.1
django-filter==23.3
drf-yasg==1.21.7
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9  # for PostgreSQL alternative
